Small changes to syntax

+++ game.cpp
@@ -1,7 +1,7 @@
 #include "game.h"
 
 Game::Game(std::string title, int width, int height)
-    :graphics{title, width, height}, dt{1.0/60.0}, lag{0.0}, performance_frequency({SDL_GetPerformanceFrequency()}),     prev_counter{SDL_GetPerformanceCounter()} {
+    :graphics{title, width, height}, dt{1.0/60.0}, lag{0.0}, performance_frequency({SDL_GetPerformanceFrequency()}), prev_counter{SDL_GetPerformanceCounter()} {
 
     // load the first "level"
     world.add_platform(0, 656, 1280, 64);
@@ -19,7 +19,7 @@ void Game::input() {
 
 void Game::update() {
     Uint64 now = SDL_GetPerformanceCounter();
-    lag += (now - prev_counter) / (double)performance_frequency;
+    lag += (now - prev_counter) / (float)performance_frequency;
     prev_counter = now;
     while (lag >= dt) {
         world.update(dt);
